Update README.md
The project was built with modern web technologies for a fast and responsive user experience. ReactJS powers the frontend, offering a dynamic, single-page app experience. Tailwind CSS was used to ensure a responsive, mobile-first design. On the backend, Node.js powers the API, and TypeScript ensures type safety throughout the project. Docker was used for containerization, allowing for seamless deployment and scalability.